converts rest of handlers to own file
