+++ lib/cell/common.js
@@ -3,7 +3,7 @@
 const { $updateCell } = require('../state');
 const Common = require('../common');
 const CellElement = require('./elements');
-const CellListeners = require('./eventListeners');
+const Prehandler = require('../prehandler');
 const CellStateUpdate = require('./stateUpdate');
 const WindowStateUpdate = require('../window/stateUpdate');
 const Styles = require('./styles');
@@ -101,7 +101,7 @@ exports.clearCell = (state, cell) => {
         value: ' '
     });
 
-    CellListeners.cellInputListener(state, cell);
+    Prehandler.cellInput(state, cell);
 };
