+++ lib/eventListener.js
@@ -1,8 +1,18 @@
 'use strict';
 
-const WindowCommon = require('./common');
-const { _state } = require('../state');
-const Handler = require('../eventHandler');
+const { _state } = require('./state');
+const Router = require('./router').router;
+
+const internals = {};
+
+internals.getCell = (state, e) => {
+
+    const input = e.target.id;
+    if (!input) return;
+    const cell = state.allCells[input.slice(5)]; // input id are prefaced with 'cell-' to slice first 5 char
+    if (!cell) return;
+    return cell;
+};
 
 
 window.addEventListener('mouseup', (e) => windowMouseup());
@@ -19,7 +29,7 @@ const windowClick = (e) => {
 
     if (e.target.nodeName === 'BUTTON') {
 
-        return new Handler({
+        return new Router({
             type: 'buttonClick',
             e
         });
@@ -29,11 +39,11 @@ const windowClick = (e) => {
 
 const windowInput = (e) => {
 
-    const cell = WindowCommon.getCell(_state, e);
+    const cell = internals.getCell(_state, e);
 
     if (!cell) return;
 
-    return new Handler({
+    return new Router({
         type: 'cellInput',
         cell
     });
@@ -42,11 +52,11 @@ const windowInput = (e) => {
 
 const windowMousedown = (e) => {
 
-    const cell = WindowCommon.getCell(_state, e);
+    const cell = internals.getCell(_state, e);
     
     if (!cell) return;
 
-    return new Handler({
+    return new Router({
         type: 'cellMousedown',
         cell
     });
@@ -55,11 +65,11 @@ const windowMousedown = (e) => {
 
 const windowMouseover = (e) => {
 
-    const cell = WindowCommon.getCell(_state, e);
+    const cell = internals.getCell(_state, e);
 
     if (!cell || !_state.mousedown) return;
 
-    return new Handler({
+    return new Router({
         type: 'cellMouseover',
         cell
     });
@@ -70,7 +80,7 @@ const windowMousemove = (e) => {
 
     if (_state.colDrag) {
 
-        return new Handler({
+        return new Router({
             type: 'resizeRowColumn',
             value: 'column',
             e
@@ -78,7 +88,7 @@ const windowMousemove = (e) => {
     }
     else if (_state.rowDrag) {
 
-        return new Handler({
+        return new Router({
             type: 'resizeRowColumn',
             value: 'row',
             e
@@ -94,7 +104,7 @@ const windowKeydown = (e) => {
         e.key === 'ArrowUp' ||
         e.key === 'ArrowDown') {
 
-        return new Handler({
+        return new Router({
             type: 'navigateCells',
             e
         });
@@ -102,14 +112,14 @@ const windowKeydown = (e) => {
 
     else if (e.key === 'Meta') {
 
-        return new Handler({
+        return new Router({
             type: 'enableCommandActive'
         });
     }
 
     if (_state.commandActive) {
 
-        return new Handler({
+        return new Router({
             type: 'commandActiveKeydown',
             e
         });
@@ -119,11 +129,11 @@ const windowKeydown = (e) => {
 
 const windowMouseup = () => {
 
-    new Handler({ type: 'windowMouseup' }
-});
+    new Router({ type: 'windowMouseup' });
+};
 
 
 const windowKeyup = () => {
 
-    new Handler({ type: 'windowKeyup' }
-});
+    new Router({ type: 'windowKeyup' });
+};
