+++ lib/handlers/clickDelete.js
@@ -1,0 +1,12 @@
+'use strict';
+
+const { $updateCell } = require('../state');
+
+
+module.exports = (state) => {
+
+    state.activeCells.forEach((id) => {
+
+        $updateCell(id, { value: ' ' })
+    });
+};
