+++ lib/handlers/index.js
@@ -1,0 +1,17 @@
+'use strict';
+
+exports.cellDrag = require('./celldrag');
+exports.cellMousedown = require('./cellMousedown');
+exports.clickCss = require('./clickCss');
+exports.clickCutCopy = require('./clickCutCopy');
+exports.clickDelete = require('./clickDelete');
+exports.clickPaste = require('./clickPaste');
+exports.clickSum = require('./clickSum');
+exports.commandActiveKeydown = require('./commandActiveKeydown');
+exports.enableCommandActive = require('./enableCommandActive');
+exports.funcCellInput = require('./funcCellInput');
+exports.funcCellOutput = require('./funcCellOutput');
+exports.keyup = require('./keyup');
+exports.mouseup = require('./mouseup');
+exports.navigateCells = require('./navigateCells');
+exports.resizeRowColumn = require('./resizeRowColumn');
