+++ lib/index.js
@@ -53,6 +53,6 @@ process.nextTick(() => {
 });
 
 main.appendChild(new DraggableDiv());
-require('./window/eventListeners');
+require('./eventListener');
