+++ lib/toolbar/elements.js
@@ -1,8 +1,6 @@
 'use strict';
 
 const Styles = require('./styles');
-const ToolbarListeners = require('./eventListeners');
-
 
 function Toolbar () {
